// src/redux/types/outputType.ts
//---------------Output Povider List---------------
export enum outputProviderActionType {
  OUTPUTPROVIDER_SUCCESS = 'OUTPUTPROVIDER_SUCCESS',
  OUTPUTPROVIDER_FAIL = 'OUTPUTPROVIDER_FAIL'
}

//---------------Output Clint List---------------
export enum outputClientActionType {
  OUTPUTCLIENTE_SUCCESS = 'OUTPUTCLIENTE_SUCCESS',
  OUTPUTCLIENTE_FAIL = 'OUTPUTCLIENTE_FAIL'
}

//---------------Output Send---------------
export enum outputSendActionType {
  OUTPUTSEND_SUCCESS = 'OUTPUTSEND_SUCCESS',
  OUTPUTSEND_FAIL = 'OUTPUTSEND_FAIL'
}
//---------------Output List---------------
export enum outputListActionType {
  OUTPUTLIST_SUCCESS = 'OUTPUTLIST_SUCCESS',
  OUTPUTLIST_FAIL = 'OUTPUTLIST_FAIL'
}
//---------------Input Inventory---------------
export enum inputInventoryActionType {
  INPUTInventory_SUCCESS = 'INPUTInventory_SUCCESS',
  INPUTInventory_FAIL = 'INPUTInventory_FAIL'
}